New package: helios4-uboot-2018.11
das U-Boot loader for the helios4 board

Builds an image for the sd card. to install to the sd card use
`dd if=/boot/u-boot-spl-sdcard.kwb of=/dev/mmcblk0 bs=512 seek=1 status=noxfer`

uboot is configured to automatically seach for an uboot script called
boot.scr and execute it. the provided boot.scr loads the kernel,
optionally with an initrd, optionally grants the booted system access to
the onboard spi nor flash and optionally enables the emmc fix. it is
configured in /boot/uboot-env.txt

There are newer versions of the kobol uboot fork, but they are mainly
targeted a different board.

see https://wiki.kobol.io/helios4/uboot/ for more information.