New package: helios4-uboot-2018.11
das U-Boot loader for the helios4 board

builds images for sd card and spi nor flash. to install to the sd card use
`dd if=/boot/u-boot-spl-sdcard.kwb of=/dev/mmcblk0 bs=512 seek=1status=noxfer`
to install to the spi nor flash uncomment `spi_workaround "on"` in
/boot/uboot-env.txt, reboot and use
`dd if=/boot/u-boot-spl-spinorflash.kwb of=/dev/mtdblock0`.

uboot is configured to automatically seach for an uboot script called
boot.scr and execute it. the provided boot.scr loads the kernel,
optionally with an initrd, optionally grants the booted system access to
the onboard spi nor flash and optionally enables the emmc fix. it is
configured in /boot/uboot-env.txt

there are newer versions of the kobol uboot fork, but this compiles for
sdcard and spi nor flash while staying in the size constraints.

see https://wiki.kobol.io/helios4/uboot/ and
https://wiki.kobol.io/helios4/spi/#write-u-boot-to-spi-nor-flash for
more information.